+++ Cargo.toml
@@ -19,12 +19,12 @@ members = [
   "examples/js_callback",
   "examples/keyed_list",
   "examples/mount_point",
-  "examples/multi_thread",
+  "examples/agents",
   "examples/nested_list",
   "examples/node_refs",
   "examples/password_strength",
   "examples/portals",
-  "examples/pub_sub",
+  "examples/contexts",
   "examples/router",
   "examples/timer",
   "examples/todomvc",
