+++ examples/agents/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "multi_thread"
+name = "agents"
 version = "0.1.0"
 authors = ["Denis Kolodin <deniskolodin@gmail.com>"]
 edition = "2018"
