+++ examples/agents/src/bin/worker.rs
