+++ examples/agents/src/lib.rs
