+++ examples/agents/src/native_worker.rs
