+++ examples/contexts/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "pub_sub"
+name = "contexts"
 version = "0.1.0"
 edition = "2018"
 license = "MIT OR Apache-2.0"
