+++ examples/contexts/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <title>Yew • Pub Sub</title>
+    <title>Yew • Context</title>
   </head>
 
   <body></body>
