+++ examples/contexts/src/msg_ctx.rs
@@ -1,0 +1,37 @@
+use std::rc::Rc;
+
+use yew::prelude::*;
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct Message {
+    pub inner: String,
+}
+
+impl Reducible for Message {
+    type Action = String;
+
+    fn reduce(self: Rc<Self>, action: Self::Action) -> Rc<Self> {
+        Message { inner: action }.into()
+    }
+}
+
+pub type MessageContext = UseReducerHandle<Message>;
+
+#[derive(Properties, Debug, PartialEq)]
+pub struct MessageProviderProps {
+    #[prop_or_default]
+    pub children: Children,
+}
+
+#[function_component]
+pub fn MessageProvider(props: &MessageProviderProps) -> Html {
+    let msg = use_reducer(|| Message {
+        inner: "No message yet.".to_string(),
+    });
+
+    html! {
+        <ContextProvider<MessageContext> context={msg}>
+            {props.children.clone()}
+        </ContextProvider<MessageContext>>
+    }
+}
