+++ examples/contexts/src/producer.rs
