+++ examples/contexts/src/subscriber.rs
