--- examples/multi_thread/README.md
@@ -1,18 +1,0 @@
-# Warning
-
-The multi-thread example is a conceptual WIP and is currently blocked on [a future Trunk feature](https://github.com/thedodd/trunk/issues/46)
-
-There is an alternate multi-thread example [here](https://github.com/yewstack/yew/tree/master/examples/web_worker_fib).
-
-
-# Multi-Thread Example
-
-[![Demo](https://img.shields.io/website?label=demo&url=https%3A%2F%2Fexamples.yew.rs%2Fmulti_thread)](https://examples.yew.rs/multi_thread)
-
-
-## Concepts
-
-Uses an [Agent] that runs in a [Web Worker].
-
-[agent]: https://yew.rs/docs/concepts/agents/
-[web worker]: https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Using_web_workers
